Added CGNS mesh reading capabilities
A mesh reading class has been added for reading CGNS mesh files,
currently only reading and processing 2D meshes. Extension to 3D should
be faily straight forward. The API exposes 4 public functions to read
coordinates, their connectivity table, boundary conditions and its
connectivity information. This class does not store any of this
information and is intended to be a lightweight class (i.e. only storing
integers like number of cells and the dimension of the mesh). It is the
responsibility of the calling class to store the output.