Added a computational mesh class
This class manages the state of the computational mesh by storing all
required information that is needed during the solving stage in terms of
grid metrics such as the coordinates, connectivity table, boundary
conditions, etc. It does not implement any logic and delegates that to
other classes that directly feed into this class.