Added the parameter reading class
This class allows parameters to be read from an input file that needs to
be specified. The input file needs to be in JSON format. The parameter
handling class only exposes one static method to read from the input
file and is templated over the return type which will be used to perform
a static cast. A default value must be provided when reading a parameter
but the default value can be set to std::nullopt (requires the
<optional> header). Each parameter that is required at runtime should
be retrieved using this method and then stored for later processing,
such as the mesh file by the mesh reading class.

changed version number